Refatorei parte do código. Passei endpoints para controller, faltou passar funções para postBusiness e testar.
